pyproject.toml
[project]
name = "requirements-analysis-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "Requirements Analysis Team", email = "contact@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langchain>=0.3.25,<0.4.0",
    "langgraph>=0.4.7,<0.5.0",
    "langchain-ibm>=0.3.11,<0.4.0",
    "ibm-watsonx-ai>=1.3.20,<2.0.0",
    "streamlit>=1.45.1,<2.0.0",
    "chromadb>=1.0.10,<2.0.0",
    "pypdf>=5.5.0,<6.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "langchain-community>=0.3.24,<0.4.0",
    "nltk>=3.9.1,<4.0.0",
    "transformers>=4.52.3,<5.0.0",
    "langchain-chroma>=0.2.4,<0.3.0",
    "unstructured[local-inference]>=0.17.2,<0.18.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "requests>=2.32.0,<3.0.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
